split Attack into submodule
